% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bin}
\alias{bin}
\title{Image binaryzation.}
\usage{
bin(img, s)
}
\arguments{
\item{img}{the original image matrix}

\item{s}{the threshold value}
}
\value{
the processed image matrix
}
\description{
Image binaryzation.
}
