% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binthin.R
\name{binthin}
\alias{binthin}
\title{Convert a gray-scale image into a binary one, then thinning is performed on it. Print 3 images horizontally for comparison.}
\usage{
binthin(img, s, flag = 0)
}
\arguments{
\item{img}{the original image matrix}

\item{s}{parameter for function bin}

\item{flag}{parameter for function thin}
}
\value{
the processed image matrix
}
\description{
Convert a gray-scale image into a binary one, then thinning is performed on it. Print 3 images horizontally for comparison.
}
