% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{field_esti}
\alias{field_esti}
\title{Orientation field estimation.}
\usage{
field_esti(img, w = 16L)
}
\arguments{
\item{img}{the original image matrix}

\item{w}{the size of blocks into which the image is divided}
}
\value{
the orientation of each pixel in the image matrix
}
\description{
Orientation field estimation.
}
