% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gabor}
\alias{gabor}
\title{The function of gabor is to create a gabor filter, which is a popular linear filter mainly used for edge detection.}
\usage{
gabor(img, w = 16L, s = 5L, gsize = 16L, f = 0.1, vx = 4L, vy = 4L)
}
\arguments{
\item{img}{a matrix that a gray-scale image turns into. The value of each element is a real between 0 and 1.}

\item{w}{an integer. It indicates the size of blocks that the image is divided into. The default is 16.}

\item{s}{an integer. It indicates the size chosen for the low-pass Gaussian filter. The default is 5.}

\item{gsize}{an integer. It stands for the size of gabor filter. The default is 16.}

\item{f}{a decimal. It indicates the ridge frequency. The default is 0.1.}

\item{vx}{an integer. It indicates the standard deviations of the Gaussian envelope of the first index. The default is 4.}

\item{vy}{an integer. It indicates the standard deviations of the Gaussian envelope of the second index. The default is 4.}
}
\value{
a matrix. The shape of the matrix is same as that of the picture provided. Each value of the element is a real between 0 and 1.
}
\description{
The function of gabor is to create a gabor filter, which is a popular linear filter mainly used for edge detection.
}
