\name{IEBF-package}
\alias{IEBF-package}
\alias{IEBF}
\docType{package}
\title{
\packageTitle{IEBF}
}
\description{
The package is initially dedicated to blurred fingerprint image enhancement. Currently, three filters are taken into consideration, namely, Gabor filter, Laplace filter and Volterra filters. Additionally, evaluation method is provided.
}
\author{
Yunlin LIU liuyunlin@ruc.edu.cn; Yifan XIE xieyifan0922@ruc.edu.cn; Yao DONG dongyao@ruc.edu.cn

Maintainer: Yunlin LIU liuyunlin@ruc.edu.cn; Yifan XIE xieyifan0922@ruc.edu.cn; Yao DONG dongyao@ruc.edu.cn
}
\section{Acknowledgement}{
The work was technically supported by Professor Wei MA and Fuyi TU.
}
\references{
Arif A , Li T , Cheng C H . Blurred fingerprint image enhancement: algorithm analysis and performance evaluation[J]. Signal Image and Video Processing, 2017.
Hong L , Wan Y , Jain A . Fingerprint image enhancement: algorithm and performance evaluation[J]. IEEE Transactions on Pattern Analysis & Machine Intelligence, 1998, 20.

}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
