% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{laplace}
\alias{laplace}
\title{Image The function of laplace is to create a laplace filter, which is a linear high-pass filter mainly used for edge enhancement.}
\usage{
laplace(img, alpha = 1)
}
\arguments{
\item{img}{a matrix that a gray-scale image turns into. The value of each element is a real between 0 and 1.}

\item{alpha}{a number. It is a coefficient that decides on how laplacian filter would work on the original image. The default is 1.}
}
\value{
a matrix. The shape of the matrix is same as that of the picture provided. Each value of the element is a real between 0 and 1.
}
\description{
Image The function of laplace is to create a laplace filter, which is a linear high-pass filter mainly used for edge enhancement.
}
