% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{minutiae_set}
\alias{minutiae_set}
\title{Minutiae extract.}
\usage{
minutiae_set(img)
}
\arguments{
\item{img}{the original image matrix}
}
\value{
the Minutiae matrix, 1 if ridge ending & bifurcation, else 0
}
\description{
Minutiae extract.
}
