% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nlm}
\alias{nlm}
\title{Compute image denoising.}
\usage{
nlm(img, N, K, sigma)
}
\arguments{
\item{img}{the original image matrix}

\item{N}{the dimension of texture replacement search region}

\item{sigma}{the smoothing parameters}

\item{k}{the dimension of the texture region of the pixel to be processed}
}
\value{
the denoised image
}
\description{
Compute image denoising.
}
