% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{normal}
\alias{normal}
\title{Normalization of original pictures.}
\usage{
normal(img, m = 100, var = 100)
}
\arguments{
\item{img}{the original image matrix}

\item{m}{a number. It indicates the estimated mean of the image. The default is 100.}

\item{var}{a number. It indicates the estimated variance of the image. The default is 100.}
}
\value{
the processed image matrix. The shape of the matrix is same as that of the picture provided. Each value of the element is a real between 0 and 1.
}
\description{
Normalization of original pictures.
}
