% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pic_con.R
\name{pic_con}
\alias{pic_con}
\title{Horizontal 1, 2 or 3 images for comparison.}
\usage{
pic_con(img1, ...)
}
\arguments{
\item{img1, img2, ...}{the original image matrix}
}
\value{
none
}
\description{
Horizontal 1, 2 or 3 images for comparison.
}
