% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simi_score}
\alias{simi_score}
\title{Similarity score of two image matrixes.}
\usage{
simi_score(i1, i2, r0 = 8L, theta0 = 7L)
}
\arguments{
\item{i1}{image matrix 1}

\item{i2}{image matrix 2}

\item{r0}{the threshold value for position}

\item{theta0}{the threshold value for direction}
}
\value{
the similarity score
}
\description{
Similarity score of two image matrixes.
}
