% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{thin}
\alias{thin}
\title{Image thinning.}
\usage{
thin(img, flag)
}
\arguments{
\item{img}{the original image matrix, which is expected to represent a binary image}

\item{flag}{the pixel value of foreground objects, 0 for black, 1 for white}
}
\value{
the processed image matrix
}
\description{
Image thinning.
}
