% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{volterra}
\alias{volterra}
\title{The function of volterra is to create a Volterra filter, which is a non-linear filter whose input-output relation is a Volterra series and it represents the most natural extension of linear filters.}
\usage{
volterra(img, c1 = 3, c2 = -0.5, c3 = -0.5, c4 = -1, c5 = -1, rho = 0.5)
}
\arguments{
\item{img}{a matrix that a gray-scale image turns into. The value of each element is a real between 0 and 1.}

\item{c1}{a number. Eq: y(n1,n2) = C1 · x2(n1,n2) − C2 · x(n1+1, n2+1) · x (n1−1, n2−1) − C3 ·x(n1+1, n2−1) · x(n1−1, n2+1) − C4 · x(n1+1, n2) · x (n1-1, n2) − C5 · x(n1, n2+1) · x(n1, n2−1) It is the first coefficient of the equation above. The default is 3.}

\item{c2}{a number. It is the second coefficient of the equation above. The default is -0.5.}

\item{c3}{a number. It is the third coefficient of the equation above. The default is -0.5.}

\item{c4}{a number. It is the fourth coefficient of the equation above. The default is -1.}

\item{c5}{a number. It is the fifth coefficient of the equation above. The default is -1.}

\item{rho}{A number. It is a coefficient that decides on how volterra filter would work on the original image. The default is 0.5.}
}
\value{
a matrix. The shape of the matrix is same as that of the picture provided. Each value of the element is a real between 0 and 1.
}
\description{
The function of volterra is to create a Volterra filter, which is a non-linear filter whose input-output relation is a Volterra series and it represents the most natural extension of linear filters.
}
